Seems that some things were missing in the last commit. Magincia Decoration Items, MOTD archive directory, ect.
Some files show that the only thing that has changed is a } at the end of the file. Why? I have no clue.

Things I actually changed were as follows:
Added additional crops so that players can grow their own regs.
grapes, pumpkin, mandrakeroot, nightshade, bloodmoss, ginsing.

Adjusted the timer in the crops so that there is only a 1 second delay between harvesting (double clicking on the plant) to get the regs.

Added the new crops to the UpRootGump, so that when the player clicks "OK to remove the plant" it will now actually remove the plant.

Changed some of the crops to give more regs depending on the skill they have. For example a Cooking skilled player would get more resources out of food crop then a mage skilled player would, and a mage skilled player would get more out of a Mandrakeroot crop then a cook skilled player would.

Fixed the MainPlantGump.cs file so that when a player adds water using the gump water pitcher, it will reopen the gump automatically instead of just dissapearing.

Adjusted DefTailoring.cs for the Skill level requirements on crafting the bag, backpack, pouch, and blank scrolls.
same with the DefCarpentry.cs file for the blank scrolls.

Compile.bat added pause in the batch file. I thought it was there originaly.

Still don't upload Log files, Saves, ect., due to security reasons.