Added: PlayerConsoleLogin - Sends message to console as to which char the account has loged onto. AnimationLocator - GM-Tool. To help find animations of creatures. InternalItems - GM-Tool. To clean items off of internal map. PlayersWealth - To get a html file placed in the root of the server with list of all the players wealth. Milk and Cheese craft ChickenEggLayer DragonKnights - Good Npc/Mobs who fight evil. Still need work on Stats. ForgottenOne - Evil Npc/Mobs who fight anything. Still need work on stats. Raven - NPC deco.
Edited:
AccountHandler - Added code so that it will show a time when incident happens.
SBFarmhand - Added Milk and cheese items to the list of selling.
LoginStats - Added auto-command to boost speed of Gms on login.
SBAnimalTrainer - Tried to straiten out the list and added Egg laying chicken to the list.
SBBaker- Added Milk and cheese items to the list of selling.
SBCook- Added Milk and cheese items to the list of selling.
MessagePump - Added code so that it will show a time when incident happens.
NetState - Added code so that it will show a time when incident happens.


Removed:
Testing Directory.

Sorted:
Mobs and Npc into directories.